XRDDEV-39 Review fixes
